refactor: rename racingCar to carRacingGame
